moved builder out of Definition namespace
